refactor: Moved Tests of TDPDataMatrix and removed unnecessary imports
