+++ plugins/nf-co2footprint/src/test/nextflow/co2footprint/utils/DataContainersTest.groovy
@@ -182,139 +182,3 @@ class DataMatrixTest extends  Specification {
         df == df2
     }
 }
-
-@Stepwise
-class TDPDataMatrixTest extends Specification {
-    private final Path tempPath = Files.createTempDirectory('tmpdir')
-    final Matrix df = new TDPDataMatrix(
-            [
-                    [13, 1, 1],
-                    [42, 10**3, 2*10**3],
-                    [100, 4, 8],
-            ],
-            ['tdp (W)', 'cores', 'threads'] as LinkedHashSet,
-            ['Intel i3-Fantasy', 'AMD YPS-x42', 'default'] as LinkedHashSet,
-    )
-    Logger logger = (Logger) LoggerFactory.getLogger(TDPDataMatrix)
-    ListAppender<ILoggingEvent> listAppender = new ListAppender<>()
-
-    def 'Should get a valid DataMatrix Extension' () {
-        setup:
-        Path filePath = tempPath.resolve('tdpDM.csv')
-        df.saveCsv(filePath)
-
-        when:
-        Matrix tdpDM = TDPDataMatrix.loadCsv(filePath, ',', 0, 0)
-
-        then:
-        tdpDM == df
-    }
-
-
-    def 'Should get a valid DataMatrix Extension' () {
-        setup:
-        String modelName = 'AMD YPS-x42'
-        Matrix model = new TDPDataMatrix(
-            [
-                [42, 10**3, 2*10**3],
-            ],
-            ['tdp (W)', 'cores', 'threads'] as LinkedHashSet,
-            [modelName] as LinkedHashSet,
-        )
-
-        when:
-        DataMatrix modelRes = df.matchModel(modelName)
-
-        then:
-        modelRes == model
-        modelRes.get(0, "tdp (W)", true) == 42
-    }
-
-    def 'Should return correct TPD, Core & Thread values' () {
-        setup:
-        Matrix df2 = new TDPDataMatrix(
-                df.data, df.getOrderedColumnKeys(), df.getOrderedRowKeys(),
-                'default', 20, 2, 4
-        )
-
-        expect:
-        df.getTDP() == 13.0
-        df.getCores() == 1
-        df.getThreads() == 1
-        df.getTDP(null, 'default') == 100
-        df2.getTDP() == 20.0
-        df2.getCores() == 2
-        df2.getThreads() == 4
-    }
-
-    def 'Should return first name of cpu model' () {
-        when:
-        Object firstName = df.getFirstName()
-
-        then:
-        firstName == 'Intel i3-Fantasy'
-    }
-
-    def 'Should return correct TPD per Core/Thread' () {
-        setup:
-        Matrix df2 = new TDPDataMatrix(
-                df.data, df.getOrderedColumnKeys(), df.getOrderedRowKeys(),
-                'default', 20, 2, 4
-        )
-        listAppender.start()
-        logger.addAppender(listAppender)
-
-        when:
-        Double dfTDPPerCore = df.matchModel('Non-existent').getCoreTDP()
-        Double df2TDPPerCore = df2.getCoreTDP()
-        Double dfTDPPerThread = df.matchModel('Non-existent').getThreadTDP()
-        Double df2TDPPerThread = df2.getThreadTDP()
-
-        then:
-        dfTDPPerCore == 25.0
-        df2TDPPerCore == 10.0
-        dfTDPPerThread == 12.5
-        df2TDPPerThread == 5.0
-        listAppender.list[0].toString() ==  '[WARN] Could not find CPU model "Non-existent" in given TDP data table. ' +
-                                            'Using default CPU power draw value (100.0 W).'
-        listAppender.list[1].toString() ==  '[WARN] Could not find CPU model "Non-existent" in given TDP data table. ' +
-                                            'Using default CPU power draw value (100.0 W).'
-
-        cleanup:
-        listAppender.list.clear()
-        logger.detachAndStopAllAppenders()
-        listAppender.stop()
-    }
-
-    def 'Should match the model names correctly' () {
-        setup:
-        listAppender.start()
-        logger.addAppender(listAppender)
-
-        expect:
-        // Standard calls
-        df.matchModel('Intel i3-Fantasy').getData() == [[13, 1, 1]]
-        df.matchModel('AMD YPS-x42').getData() == [[42, 10**3, 2 * 10**3]]
-        df.matchModel('default').getData() == [[100, 4, 8]]
-
-        // Handling different string possibilities
-        df.matchModel('Intel® i3-Fantasy™').getData() == [[13, 1, 1]]
-        df.matchModel('Intel(R) i3-Fantasy(TM)').getData() == [[13, 1, 1]]
-        df.matchModel('Intel® i3-Fantasy(TM)').getData() == [[13, 1, 1]]
-        df.matchModel('Intel® i3-Fantasy(TM) @ 10Trillion GW').getData() == [[13, 1, 1]]
-        df.matchModel('Intel® i3-Fantasy(TM) @ 10Trillion GW @ 0.00001MHz').getData() == [[13, 1, 1]]
-
-        // Getting default when non-existent
-        df.matchModel('Non-existent2').getData() == [[100, 4, 8]]
-        listAppender.list[0].toString() == '[WARN] Could not find CPU model "Non-existent2" in given TDP data table. ' +
-                'Using default CPU power draw value (100.0 W).'
-        df.matchModel('Intel® i3-Fantasy(TM) 10Trillion GW, 0.00001MHz').getData() == [[100, 4, 8]]
-        listAppender.list[1].toString() == '[WARN] Could not find CPU model "Intel® i3-Fantasy(TM) 10Trillion GW, 0.00001MHz" in given TDP data table. ' +
-                'Using default CPU power draw value (100.0 W).'
-
-        cleanup:
-        listAppender.list.clear()
-        logger.detachAndStopAllAppenders()
-        listAppender.stop()
-    }
-}
